Migrated threshold logic to extension-side
